backend/requirements.txt
# Architectural Detail Comparison Tool - Backend Dependencies
# Python 3.11+

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# PDF rasterization
PyMuPDF>=1.24.0

# Image processing
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Utilities
python-dateutil>=2.8.0

# AI/ML (optional - for difference summarization)
google-generativeai>=0.3.0
